feat(CommandMenu): add keyboard shortcuts menu
