+++ src/livecodes/core.ts
@@ -57,8 +57,6 @@ import type {
   SDKEvent,
   Editor,
   AppLanguage,
-  INinjaAction,
-  TemplateName,
 } from './models';
 import type { GitHubFile } from './services/github';
 import type {
@@ -113,7 +111,7 @@ import {
   predefinedValues,
   colorToHex,
   capitalize,
-  stringUnionToArray,
+  isMac,
 } from './utils';
 import { compress } from './utils/compression';
 import { getCompiler, getAllCompilers, cjs2esm, getCompileResult } from './compiler';
@@ -161,6 +159,7 @@ import type {
 } from './i18n';
 import { appLanguages } from './i18n/app-languages';
 import { themeColors } from './UI/theme-colors';
+import { getCommandMenuActions } from './UI/command-menu-actions';
 
 // declare global dependencies
 declare global {
@@ -186,7 +185,7 @@ declare global {
 const stores: Stores = createStores();
 const eventsManager = createEventsManager();
 let notifications: ReturnType<typeof createNotifications>;
-let modal: ReturnType<typeof createModal>;
+export let modal: ReturnType<typeof createModal>;
 let i18n: Await<ReturnType<typeof import('./i18n').init>> | undefined;
 let split: ReturnType<typeof createSplitPanes> | null = null;
 let typeLoader: ReturnType<typeof createTypeLoader>;
@@ -203,7 +202,7 @@ let formatter: Formatter;
 let editors: Editors;
 let customEditors: CustomEditors;
 let toolsPane: ToolsPane | undefined;
-let authService: ReturnType<typeof createAuthService> | undefined;
+export let authService: ReturnType<typeof createAuthService> | undefined;
 let editorLanguages: EditorLanguages | undefined;
 let resultLanguages: Language[] = [];
 let projectId: string;
@@ -1795,7 +1794,7 @@ const setTheme = (theme: Theme, editorTheme: Config['editorTheme']) => {
     customEditors[editor?.getLanguage()]?.setTheme(theme);
   });
   toolsPane?.console?.setTheme?.(theme);
-  document.querySelector('ninja-keys')?.classList.toggle('dark', theme === 'dark');
+  UI.getNinjaKeys()?.classList.toggle('dark', theme === 'dark');
 };
 
 const changeThemeColor = () => {
@@ -2364,7 +2363,7 @@ const handleChangeContent = () => {
 
 const handleKeyboardShortcuts = () => {
   let lastkeys = '';
-  const ctrl = (e: KeyboardEvent) => (navigator.platform.match('Mac') ? e.metaKey : e.ctrlKey);
+  const ctrl = (e: KeyboardEvent) => (isMac() ? e.metaKey : e.ctrlKey);
 
   const hotKeys = async (e: KeyboardEvent) => {
     // Ctrl + P opens the command palette
@@ -2420,6 +2419,41 @@ const handleKeyboardShortcuts = () => {
       e.preventDefault();
       UI.getResultButton()?.click();
       lastkeys = 'Ctrl + Alt + R';
+      return;
+    }
+
+    // Ctrl + Alt + Z toggles result zoom
+    if (ctrl(e) && e.altKey && e.code === 'KeyZ') {
+      e.preventDefault();
+      UI.getZoomButton()?.click();
+      lastkeys = 'Ctrl + Alt + Z';
+      return;
+    }
+
+    // Ctrl + Alt + E focuses active editor
+    if (ctrl(e) && e.altKey && e.code === 'KeyE') {
+      e.preventDefault();
+      getActiveEditor().focus();
+      lastkeys = 'Ctrl + Alt + E';
+      return;
+    }
+
+    // Esc + Esc moves focus out of editor
+    // Esc + Esc + Esc moves focus to logo
+    if (e.code === 'Escape') {
+      if (lastkeys === 'Esc') {
+        e.preventDefault();
+        UI.getFocusButton()?.focus();
+        lastkeys = 'Esc + Esc';
+        return;
+      }
+      if (lastkeys === 'Esc + Esc') {
+        e.preventDefault();
+        UI.getLogoLink()?.focus();
+        lastkeys = 'Esc + Esc + Esc';
+        return;
+      }
+      lastkeys = 'Esc';
       return;
     }
 
@@ -2492,15 +2526,11 @@ const handleKeyboardShortcuts = () => {
       return;
     }
 
-    // Ctrl + K Z toggles focus mode
-    if (ctrl(e) && e.code === 'KeyK') {
-      lastkeys = 'Ctrl + K';
-      return;
-    }
-    if (ctrl(e) && e.code === 'KeyZ' && lastkeys === 'Ctrl + K') {
+    // Ctrl + Alt + F toggles focus mode
+    if (ctrl(e) && e.altKey && e.code === 'KeyF') {
       e.preventDefault();
       UI.getFocusButton()?.click();
-      lastkeys = 'Z';
+      lastkeys = 'Ctrl + Alt + F';
       return;
     }
 
@@ -2521,829 +2551,11 @@ const handleCommandMenu = async () => {
   loadStylesheet(fontMaterialIconsUrl, 'material-icons');
   await loadNinjaKeys();
 
-  const ninja = document.querySelector('ninja-keys') as any;
-
-  const changeMenuSetting = (setting: keyof Config, checked: boolean) => {
-    const toggle = [...UI.getSettingToggles()].find((t) => t.dataset.config === setting);
-    if (toggle) {
-      toggle.checked = checked;
-      toggle.dispatchEvent(new Event('change'));
-    }
-  };
-
-  const actions: INinjaAction[] = [
-    {
-      id: 'Show',
-      title: window.deps.translateString('commandMenu.show.title', 'Show …'),
-      mdIcon: 'visibility',
-      children: [
-        {
-          id: 'Next Editor',
-          title: window.deps.translateString('commandMenu.show.next', 'Show Next Editor'),
-          hotkey: 'ctrl+alt+ArrowRight',
-          mdIcon: 'skip_next',
-          handler: () => {
-            document.dispatchEvent(
-              new KeyboardEvent('keydown', {
-                ctrlKey: true,
-                altKey: true,
-                key: 'ArrowRight',
-                code: 'ArrowRight',
-              }),
-            );
-          },
-        },
-        {
-          id: 'Previous Editor',
-          title: window.deps.translateString('commandMenu.show.previous', 'Show Previous Editor'),
-          hotkey: 'ctrl+alt+ArrowLeft',
-          mdIcon: 'skip_previous',
-          handler: () => {
-            document.dispatchEvent(
-              new KeyboardEvent('keydown', {
-                ctrlKey: true,
-                altKey: true,
-                key: 'ArrowLeft',
-                code: 'ArrowLeft',
-              }),
-            );
-          },
-        },
-        {
-          id: 'Markup Editor',
-          title: window.deps.translateString('commandMenu.show.markup', 'Show Markup Editor'),
-          hotkey: 'ctrl+alt+1',
-          mdIcon: 'html',
-          handler: () => {
-            UI.getMarkupEditorTitle()?.click();
-          },
-        },
-        {
-          id: 'Style Editor',
-          title: window.deps.translateString('commandMenu.show.style', 'Show Style Editor'),
-          hotkey: 'ctrl+alt+2',
-          mdIcon: 'css',
-          handler: () => {
-            UI.getStyleEditorTitle()?.click();
-          },
-        },
-        {
-          id: 'Script Editor',
-          title: window.deps.translateString('commandMenu.show.script', 'Show Script Editor'),
-          hotkey: 'ctrl+alt+3',
-          mdIcon: 'javascript',
-          handler: () => {
-            UI.getScriptEditorTitle()?.click();
-          },
-        },
-        {
-          id: 'Toggle Result',
-          title: window.deps.translateString('commandMenu.show.result', 'Toggle Result'),
-          hotkey: 'ctrl+alt+R',
-          // mdIcon: 'split_scene',
-          icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="M160-160q-33 0-56.5-23.5T80-240v-480q0-33 23.5-56.5T160-800h200v80H160v480h200v80H160Zm280 80v-800h80v80h280q33 0 56.5 23.5T880-720v480q0 33-23.5 56.5T800-160H520v80h-80Zm80-160h280v-480H520v480Zm-360 0v-480 480Zm640 0v-480 480Z"/></svg>`,
-          handler: () => {
-            UI.getResultButton()?.click();
-          },
-        },
-        {
-          id: 'Toggle Console',
-          title: window.deps.translateString('commandMenu.show.console', 'Toggle Console'),
-          hotkey: 'ctrl+alt+C',
-          mdIcon: 'terminal',
-          handler: () => {
-            UI.getConsoleButton()?.dispatchEvent(new Event('touchstart'));
-          },
-        },
-        {
-          id: 'Maximize Console',
-          title: window.deps.translateString(
-            'commandMenu.show.maximizeConsole',
-            'Maximize Console',
-          ),
-          hotkey: 'ctrl+alt+C+F',
-          mdIcon: 'terminal',
-          handler: () => {
-            UI.getConsoleButton()?.dispatchEvent(new Event('dblclick'));
-          },
-        },
-        {
-          id: 'Toggle Compiled Code',
-          title: window.deps.translateString('commandMenu.show.compiled', 'Toggle Compiled Code'),
-          // mdIcon: 'code_blocks',
-          icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="m384-336 56-57-87-87 87-87-56-57-144 144 144 144Zm192 0 144-144-144-144-56 57 87 87-87 87 56 57ZM200-120q-33 0-56.5-23.5T120-200v-560q0-33 23.5-56.5T200-840h560q33 0 56.5 23.5T840-760v560q0 33-23.5 56.5T760-120H200Zm0-80h560v-560H200v560Zm0-560v560-560Z"/></svg>`,
-          handler: () => {
-            UI.getCompiledButton()?.dispatchEvent(new Event('touchstart'));
-          },
-        },
-        {
-          id: 'Maximize Compiled Code',
-          title: window.deps.translateString(
-            'commandMenu.show.maximizeCompiled',
-            'Maximize Compiled Code',
-          ),
-          icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="m384-336 56-57-87-87 87-87-56-57-144 144 144 144Zm192 0 144-144-144-144-56 57 87 87-87 87 56 57ZM200-120q-33 0-56.5-23.5T120-200v-560q0-33 23.5-56.5T200-840h560q33 0 56.5 23.5T840-760v560q0 33-23.5 56.5T760-120H200Zm0-80h560v-560H200v560Zm0-560v560-560Z"/></svg>`,
-          handler: () => {
-            UI.getCompiledButton()?.dispatchEvent(new Event('dblclick'));
-          },
-        },
-        {
-          id: 'Toggle Tests',
-          title: window.deps.translateString('commandMenu.show.tests', 'Toggle Tests'),
-          // mdIcon: 'labs',
-          icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="M480-80q-83 0-141.5-58.5T280-280v-360q-33 0-56.5-23.5T200-720v-80q0-33 23.5-56.5T280-880h400q33 0 56.5 23.5T760-800v80q0 33-23.5 56.5T680-640v360q0 83-58.5 141.5T480-80ZM280-720h400v-80H280v80Zm200 560q50 0 85-35t35-85H480v-80h120v-80H480v-80h120v-120H360v360q0 50 35 85t85 35ZM280-720v-80 80Z"/></svg>`,
-          handler: () => {
-            UI.getTestsButton()?.dispatchEvent(new Event('touchstart'));
-          },
-        },
-        {
-          id: 'Maximize Tests',
-          title: window.deps.translateString('commandMenu.show.maximizeTests', 'Maximize Tests'),
-          // mdIcon: 'labs',
-          icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="M480-80q-83 0-141.5-58.5T280-280v-360q-33 0-56.5-23.5T200-720v-80q0-33 23.5-56.5T280-880h400q33 0 56.5 23.5T760-800v80q0 33-23.5 56.5T680-640v360q0 83-58.5 141.5T480-80ZM280-720h400v-80H280v80Zm200 560q50 0 85-35t35-85H480v-80h120v-80H480v-80h120v-120H360v360q0 50 35 85t85 35ZM280-720v-80 80Z"/></svg>`,
-          handler: () => {
-            UI.getTestsButton()?.dispatchEvent(new Event('dblclick'));
-          },
-        },
-        {
-          id: 'Toggle Full Screen',
-          title: window.deps.translateString('commandMenu.show.fullscreen', 'Toggle Full Screen'),
-          hotkey: 'f11',
-          mdIcon: 'zoom_out_map',
-          handler: () => {
-            UI.getFullscreenButton()?.click();
-          },
-        },
-        {
-          id: 'Toggle Focus Mode',
-          title: window.deps.translateString('commandMenu.show.focusMode', 'Toggle Focus Mode'),
-          hotkey: 'ctrl+K+Z',
-          mdIcon: 'crop_free',
-          handler: () => {
-            UI.getFocusButton()?.click();
-          },
-        },
-      ],
-    },
-    {
-      id: 'Select Language',
-      title: window.deps.translateString('commandMenu.selectLanguage', 'Select Language'),
-      mdIcon: 'code',
-      children: languages
-        .sort((a, b) => a.title.localeCompare(b.title))
-        .map((lang) => ({
-          id: 'Language: ' + lang.title,
-          title: lang.longTitle ?? lang.title,
-          keywords: [lang.name, lang.title, lang.longTitle, ...lang.extensions].join(', '),
-          handler: async () => {
-            document
-              .querySelector<HTMLAnchorElement>('a[data-editor][data-lang="' + lang.name + '"]')
-              ?.dispatchEvent(new Event('mousedown'));
-          },
-        })),
-    },
-    {
-      id: 'Starter Templates',
-      title: window.deps.translateString('commandMenu.starterTemplates', 'Starter Templates'),
-      mdIcon: 'library_books',
-      children: stringUnionToArray<TemplateName>()(
-        'blank',
-        'javascript',
-        'typescript',
-        'react',
-        'react-native',
-        'vue2',
-        'vue',
-        'angular',
-        'preact',
-        'svelte',
-        'solid',
-        'lit',
-        'stencil',
-        'mdx',
-        'astro',
-        'riot',
-        'malina',
-        'jquery',
-        'backbone',
-        'knockout',
-        'jest',
-        'jest-react',
-        'bootstrap',
-        'tailwindcss',
-        'd3',
-        'phaser',
-        'coffeescript',
-        'livescript',
-        'civet',
-        'clio',
-        'imba',
-        'rescript',
-        'reason',
-        'ocaml',
-        'python',
-        'pyodide',
-        'python-wasm',
-        'r',
-        'ruby',
-        'ruby-wasm',
-        'go',
-        'php',
-        'php-wasm',
-        'cpp',
-        'clang',
-        'cpp-wasm',
-        'perl',
-        'lua',
-        'lua-wasm',
-        'teal',
-        'fennel',
-        'julia',
-        'scheme',
-        'commonlisp',
-        'clojurescript',
-        'gleam',
-        'tcl',
-        'markdown',
-        'assemblyscript',
-        'wat',
-        'sql',
-        'postgresql',
-        'prolog',
-        'blockly',
-        'diagrams',
-      ).map((template) => ({
-        id: 'Starter template: ' + template,
-        title: window.deps.translateString('commandMenu.template', 'Template') + ': ' + template,
-        handler: async () => {
-          await loadStarterTemplate(template);
-        },
-      })),
-    },
-    {
-      id: 'Run',
-      title: window.deps.translateString('commandMenu.run', 'Run'),
-      hotkey: 'shift+Enter',
-      mdIcon: 'play_arrow',
-      handler: () => {
-        UI.getRunButton()?.click();
-      },
-    },
-    {
-      id: 'Share',
-      title: window.deps.translateString('menu.share', 'Share …'),
-      hotkey: 'ctrl+alt+S',
-      mdIcon: 'share',
-      handler: () => {
-        UI.getShareLink()?.click();
-      },
-    },
-    {
-      id: 'New',
-      title: window.deps.translateString('menu.new', 'New …'),
-      hotkey: 'ctrl+alt+N',
-      mdIcon: 'note_add',
-      handler: () => {
-        UI.getNewLink()?.click();
-      },
-    },
-    {
-      id: 'Open',
-      title: window.deps.translateString('menu.open', 'Open …'),
-      hotkey: 'ctrl+O',
-      mdIcon: 'file_open',
-      handler: () => {
-        UI.getOpenLink()?.click();
-      },
-    },
-    {
-      id: 'Save',
-      title: window.deps.translateString('menu.save', 'Save'),
-      hotkey: 'ctrl+S',
-      mdIcon: 'save',
-      handler: () => {
-        UI.getSaveLink()?.click();
-      },
-    },
-    {
-      id: 'Save As',
-      title: window.deps.translateString('menu.saveAs.heading', 'Save as …'),
-      mdIcon: 'save_as',
-      children: [
-        {
-          id: 'Save as a fork',
-          title: window.deps.translateString(
-            'commandMenu.saveAsFork',
-            'Save as a Fork (New Project)',
-          ),
-          hotkey: 'ctrl+shift+S',
-          mdIcon: 'save_as',
-          handler: () => {
-            UI.getForkLink()?.click();
-          },
-        },
-        {
-          id: 'Save as a template',
-          title: window.deps.translateString('commandMenu.saveAsTemplate', 'Save as a Template'),
-          mdIcon: 'library_add',
-          handler: async () => {
-            UI.getSaveAsTemplateLink()?.click();
-          },
-        },
-      ],
-    },
-    {
-      id: 'Import',
-      title: window.deps.translateString('menu.import', 'Import …'),
-      hotkey: 'ctrl+alt+I',
-      mdIcon: 'upload',
-      handler: () => {
-        UI.getImportLink()?.click();
-      },
-    },
-    {
-      id: 'Export',
-      title: window.deps.translateString('menu.export.heading', 'Export'),
-      mdIcon: 'download',
-      children: [
-        {
-          id: 'Export as JSON',
-          title: window.deps.translateString('menu.export.json', 'Export Project (JSON)'),
-          mdIcon: 'data_object',
-          handler: () => {
-            UI.getExportJSONLink()?.click();
-          },
-        },
-        {
-          id: 'Export as HTML',
-          title: window.deps.translateString('menu.export.result', 'Export Result (HTML)'),
-          mdIcon: 'html',
-          handler: () => {
-            UI.getExportResultLink()?.click();
-          },
-        },
-        {
-          id: 'Export as ZIP',
-          title: window.deps.translateString('menu.export.src', 'Export Source (ZIP)'),
-          mdIcon: 'archive',
-          handler: () => {
-            UI.getExportSourceLink()?.click();
-          },
-        },
-        {
-          id: 'Export to GitHub Gist',
-          title: window.deps.translateString('menu.export.gist', 'Export to GitHub Gist'),
-          mdIcon: 'north_east',
-          handler: () => {
-            UI.getExportGithubGistLink()?.click();
-          },
-        },
-        {
-          id: 'Export to Codepen',
-          title: window.deps.translateString('menu.export.codepen', 'Edit in CodePen'),
-          mdIcon: 'north_east',
-          handler: () => {
-            UI.getExportCodepenLink()?.click();
-          },
-        },
-        {
-          id: 'Export to Fiddle',
-          title: window.deps.translateString('menu.export.jsfiddle', 'Edit in JSFiddle'),
-          mdIcon: 'north_east',
-          handler: () => {
-            UI.getExportJsfiddleLink()?.click();
-          },
-        },
-      ],
-    },
-    {
-      id: 'Deploy',
-      title: window.deps.translateString('menu.deploy', 'Deploy …'),
-      mdIcon: 'rocket_launch',
-      handler: () => {
-        UI.getDeployLink()?.click();
-      },
-    },
-    {
-      id: 'Broadcast',
-      title: window.deps.translateString('menu.broadcast', 'Broadcast …'),
-      mdIcon: 'rss_feed',
-      handler: () => {
-        UI.getBroadcastLink()?.click();
-      },
-    },
-    {
-      id: 'Embed',
-      title: window.deps.translateString('menu.embed', 'Embed …'),
-      mdIcon: 'widgets',
-      handler: () => {
-        UI.getEmbedLink()?.click();
-      },
-    },
-    {
-      id: 'Project Info',
-      title: window.deps.translateString('menu.project', 'Project Info …'),
-      mdIcon: 'info',
-      handler: () => {
-        UI.getProjectInfoLink()?.click();
-      },
-    },
-    {
-      id: 'Custom Settings',
-      title: window.deps.translateString('menu.customSettings', 'Custom Settings …'),
-      mdIcon: 'data_object',
-      handler: () => {
-        UI.getCustomSettingsLink()?.click();
-      },
-    },
-    {
-      id: 'External Resources',
-      title: window.deps.translateString('menu.resources', 'External Resources …'),
-      mdIcon: 'file_present',
-      handler: () => {
-        UI.getExternalResourcesLink()?.click();
-      },
-    },
-    {
-      id: 'Assets',
-      title: window.deps.translateString('menu.assets', 'Assets …'),
-      mdIcon: 'perm_media',
-      handler: () => {
-        UI.getAssetsLink()?.click();
-      },
-    },
-    {
-      id: 'Code Snippets',
-      title: window.deps.translateString('menu.snippets', 'Code Snippets …'),
-      mdIcon: 'text_snippet',
-      handler: () => {
-        UI.getSnippetsLink()?.click();
-      },
-    },
-    {
-      id: 'Backup / Restore',
-      title: window.deps.translateString('menu.backup', 'Backup / Restore …'),
-      // mdIcon: 'deployed_code_update',
-      icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="m720-80 120-120-28-28-72 72v-164h-40v164l-72-72-28 28L720-80ZM480-800 243-663l237 137 237-137-237-137ZM120-321v-318q0-22 10.5-40t29.5-29l280-161q10-5 19.5-8t20.5-3q11 0 21 3t19 8l280 161q19 11 29.5 29t10.5 40v159h-80v-116L479-434 200-596v274l240 139v92L160-252q-19-11-29.5-29T120-321ZM720 0q-83 0-141.5-58.5T520-200q0-83 58.5-141.5T720-400q83 0 141.5 58.5T920-200q0 83-58.5 141.5T720 0ZM480-491Z"/></svg>`,
-      handler: () => {
-        UI.getBackupLink()?.click();
-      },
-    },
-    {
-      id: 'Welcome Screen',
-      title: window.deps.translateString('menu.welcome.heading', 'Welcome …'),
-      mdIcon: 'dashboard',
-      handler: () => {
-        UI.getWelcomeLink()?.click();
-      },
-    },
-    {
-      id: 'Settings',
-      title: window.deps.translateString('menu.appSettings.heading', 'Settings'),
-      mdIcon: 'settings',
-      children: [
-        {
-          id: 'Editor Settings',
-          title: window.deps.translateString('menu.editorSettings', 'Editor Settings …'),
-          mdIcon: 'settings',
-          handler: () => {
-            UI.getEditorSettingsLink()?.click();
-          },
-        },
-        {
-          id: 'Enable AI Code Assistant',
-          title: window.deps.translateString('commandMenu.enableAI', 'Enable AI Code Assistant'),
-          mdIcon: 'settings',
-          handler: () => {
-            changeEditorSettings({ enableAI: true });
-          },
-        },
-        {
-          id: 'Disable AI Code Assistant',
-          title: window.deps.translateString('commandMenu.disableAI', 'Disable AI Code Assistant'),
-          mdIcon: 'settings',
-          handler: () => {
-            changeEditorSettings({ enableAI: false });
-          },
-        },
-        {
-          id: 'Enable Auto Update',
-          title: window.deps.translateString('commandMenu.enableAutoUpdate', 'Enable Auto Update'),
-          keywords: 'autoupdate',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('autoupdate', true);
-          },
-        },
-        {
-          id: 'Disable Auto Update',
-          title: window.deps.translateString(
-            'commandMenu.disableAutoUpdate',
-            'Disable Auto Update',
-          ),
-          keywords: 'autoupdate',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('autoupdate', false);
-          },
-        },
-        {
-          id: 'Enable Auto Save',
-          title: window.deps.translateString('commandMenu.enableAutoSave', 'Enable Auto Save'),
-          keywords: 'autosave',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('autosave', true);
-          },
-        },
-        {
-          id: 'Disable Auto Save',
-          title: window.deps.translateString('commandMenu.disableAutoSave', 'Disable Auto Save'),
-          keywords: 'autosave',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('autosave', false);
-          },
-        },
-        {
-          id: 'Enable Format On-Save',
-          title: window.deps.translateString(
-            'commandMenu.enableFormatOnSave',
-            'Enable Format On-Save',
-          ),
-          keywords: 'onsave',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('formatOnsave', true);
-          },
-        },
-        {
-          id: 'Disable Format On-Save',
-          title: window.deps.translateString(
-            'commandMenu.disableFormatOnSave',
-            'Disable Format On-Save',
-          ),
-          keywords: 'onsave',
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('formatOnsave', false);
-          },
-        },
-        {
-          id: 'Enable Recover Unsaved',
-          title: window.deps.translateString(
-            'commandMenu.enableRecoverUnsaved',
-            'Enable Recover Unsaved',
-          ),
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('recoverUnsaved', true);
-          },
-        },
-        {
-          id: 'Disable Recover Unsaved',
-          title: window.deps.translateString(
-            'commandMenu.disableRecoverUnsaved',
-            'Disable Recover Unsaved',
-          ),
-          mdIcon: 'edit',
-          handler: () => {
-            changeMenuSetting('recoverUnsaved', false);
-          },
-        },
-        {
-          id: 'Enable Vim Mode',
-          title: window.deps.translateString('commandMenu.enableVim', 'Enable Vim Mode'),
-          mdIcon: 'edit',
-          handler: () => {
-            changeEditorSettings({ editorMode: 'vim' });
-          },
-        },
-        {
-          id: 'Disable Vim Mode',
-          title: window.deps.translateString('commandMenu.disableVim', 'Disable Vim Mode'),
-          mdIcon: 'edit',
-          handler: () => {
-            changeEditorSettings({ editorMode: undefined });
-          },
-        },
-        {
-          id: 'Enable Emacs Mode',
-          title: window.deps.translateString('commandMenu.enableEmacs', 'Enable Emacs Mode'),
-          mdIcon: 'edit',
-          handler: () => {
-            changeEditorSettings({ editorMode: 'emacs' });
-          },
-        },
-        {
-          id: 'Disable Emacs Mode',
-          title: window.deps.translateString('commandMenu.disableEmacs', 'Disable Emacs Mode'),
-          mdIcon: 'edit',
-          handler: () => {
-            changeEditorSettings({ editorMode: undefined });
-          },
-        },
-        {
-          id: 'Responsive Layout',
-          title: window.deps.translateString('commandMenu.responsiveLayout', 'Responsive Layout'),
-          mdIcon: 'edit',
-          handler: () => {
-            setUserConfig({ layout: 'responsive' });
-            setLayout('responsive');
-          },
-        },
-        {
-          id: 'Vertical Layout',
-          title: window.deps.translateString('commandMenu.verticalLayout', 'Vertical Layout'),
-          mdIcon: 'edit',
-          handler: () => {
-            setUserConfig({ layout: 'vertical' });
-            setLayout('vertical');
-          },
-        },
-        {
-          id: 'Horizontal Layout',
-          title: window.deps.translateString('commandMenu.horizontalLayout', 'Horizontal Layout'),
-          mdIcon: 'edit',
-          handler: () => {
-            setUserConfig({ layout: 'horizontal' });
-            setLayout('horizontal');
-          },
-        },
-      ],
-    },
-    {
-      id: 'Format Code',
-      title: window.deps.translateString('commandMenu.formatCode', 'Format Code'),
-      hotkey: 'shift+alt+f',
-      mdIcon: 'format_align_left',
-      handler: () => {
-        UI.getFormatButton()?.click();
-      },
-    },
-    {
-      id: 'Copy Code',
-      title: window.deps.translateString('commandMenu.copy', 'Copy Code'),
-      mdIcon: 'content_copy',
-      handler: () => {
-        UI.getCopyButton()?.click();
-      },
-    },
-    {
-      id: 'Copy Code as Data URL',
-      title: window.deps.translateString('commandMenu.copyAsDataUrl', 'Copy Code as Data URL'),
-      mdIcon: 'dataset_linked',
-      handler: () => {
-        UI.getCopyAsUrlButton()?.click();
-      },
-    },
-    {
-      id: 'Code to Image',
-      title: window.deps.translateString('app.codeToImage.hint', 'Code to Image'),
-      mdIcon: 'camera',
-      handler: () => {
-        UI.getCodeToImageButton()?.click();
-      },
-    },
-    {
-      id: 'Run Tests',
-      title: window.deps.translateString('commandMenu.show.runTests', 'Run Tests'),
-      hotkey: 'ctrl+alt+T',
-      // mdIcon: 'labs',
-      icon: `<svg xmlns="http://www.w3.org/2000/svg" height="20px" viewBox="0 -960 960 960" width="20px" fill="currentColor"><path d="M480-80q-83 0-141.5-58.5T280-280v-360q-33 0-56.5-23.5T200-720v-80q0-33 23.5-56.5T280-880h400q33 0 56.5 23.5T760-800v80q0 33-23.5 56.5T680-640v360q0 83-58.5 141.5T480-80ZM280-720h400v-80H280v80Zm200 560q50 0 85-35t35-85H480v-80h120v-80H480v-80h120v-120H360v360q0 50 35 85t85 35ZM280-720v-80 80Z"/></svg>`,
-      handler: async () => {
-        UI.getRunTestsButton()?.click();
-      },
-    },
-    {
-      id: 'Change UI Language',
-      title: window.deps.translateString('commandMenu.changeUILanguage', 'Change UI Language'),
-      mdIcon: 'language',
-      children: Object.entries(appLanguages).map(([key, lang]) => ({
-        id: 'UI Language: ' + key,
-        title: lang,
-        keywords: key,
-        matcher: (
-          _action: INinjaAction,
-          { searchString }: { searchString: string; searchRegex: RegExp },
-        ) => key.includes(searchString.toLowerCase()) || lang.includes(searchString),
-        handler: async () => {
-          document
-            .querySelector<HTMLAnchorElement>('#app-menu-i18n a[data-lang="' + key + '"]')
-            ?.click();
-        },
-      })),
-    },
-    {
-      id: 'Change Theme',
-      title: window.deps.translateString('commandMenu.changeTheme.title', 'Change Theme'),
-      mdIcon: 'palette',
-      children: [
-        {
-          id: 'Light Theme',
-          title: window.deps.translateString(
-            'commandMenu.changeTheme.light',
-            'Change to Light Theme',
-          ),
-          mdIcon: 'light_mode',
-          handler: () => {
-            UI.getDarkThemeButton()?.click();
-            return { keepOpen: true };
-          },
-        },
-        {
-          id: 'Dark Theme',
-          title: window.deps.translateString(
-            'commandMenu.changeTheme.dark',
-            'Change to Dark Theme',
-          ),
-          mdIcon: 'dark_mode',
-          handler: () => {
-            UI.getLightThemeButton()?.click();
-            return { keepOpen: true };
-          },
-        },
-      ],
-    },
-    {
-      id: 'Documentation',
-      title: window.deps.translateString('menu.docs', 'Documentation'),
-      mdIcon: 'menu_book',
-      children: [
-        {
-          id: 'Documentation Home',
-          title: window.deps.translateString('menu.docs', 'Documentation'),
-          mdIcon: 'menu_book',
-          handler: () => {
-            window.open(predefinedValues.DOCS_BASE_URL, '_blank');
-          },
-        },
-        {
-          id: 'Getting Started',
-          title: window.deps.translateString('menu.getstart', 'Getting Started'),
-          mdIcon: 'menu_book',
-          handler: () => {
-            window.open(predefinedValues.DOCS_BASE_URL + 'getting-started', '_blank');
-          },
-        },
-        {
-          id: 'Features',
-          title: window.deps.translateString('menu.features', 'Features'),
-          mdIcon: 'menu_book',
-          handler: () => {
-            window.open(predefinedValues.DOCS_BASE_URL + 'features', '_blank');
-          },
-        },
-        {
-          id: 'Configuration',
-          title: window.deps.translateString('menu.config', 'Configuration'),
-          mdIcon: 'menu_book',
-          handler: () => {
-            window.open(predefinedValues.DOCS_BASE_URL + 'configuration', '_blank');
-          },
-        },
-        {
-          id: 'SDK',
-          title: window.deps.translateString('menu.sdk', 'SDK'),
-          mdIcon: 'menu_book',
-          handler: () => {
-            window.open(predefinedValues.DOCS_BASE_URL + 'sdk', '_blank');
-          },
-        },
-      ],
-    },
-    {
-      id: 'About',
-      title: window.deps.translateString('menu.about', 'About ...'),
-      mdIcon: 'contact_support',
-      handler: () => {
-        UI.getAboutLink()?.click();
-      },
-    },
-  ];
-
-  const adjustTrailingDots = (data: INinjaAction[]): INinjaAction[] =>
-    data.map((action) => ({
-      ...action,
-      children:
-        action.children?.length && action.children.every((a) => typeof a !== 'string')
-          ? adjustTrailingDots(action.children)
-          : action.children,
-      title: !action.children?.length
-        ? action.title.replace(' …', '').replace(' ...', '')
-        : action.title.endsWith(' …') || action.title.endsWith(' ...')
-          ? action.title
-          : action.title + ' …',
-    }));
+  const ninja = UI.getNinjaKeys() as any;
+  if (!ninja) return;
 
   const header = ninja.shadowRoot.querySelector('ninja-header');
-  const breadcrumb = header?.shadowRoot.querySelector('.breadcrumb-list .breadcrumb');
-  if (breadcrumb) {
-    breadcrumb.innerHTML = breadcrumb.innerHTML.replace(
-      'Home',
-      window.deps.translateString('commandMenu.home', 'Home'),
-    );
-  }
+  const HomeBreadcrumb = header?.shadowRoot.querySelector('.breadcrumb-list .breadcrumb');
 
   const closeBtn = header?.shadowRoot.querySelector('.breadcrumb-list .breadcrumb--close');
   if (closeBtn) {
@@ -3362,34 +2574,45 @@ const handleCommandMenu = async () => {
       );
   }
 
-  const openCommandMenu = () => {
+  const changeLayout = (layout: Config['layout']) => {
+    setUserConfig({ layout });
+    setLayout(layout);
+  };
+
+  const openCommandMenu = (parent?: string) => {
     modal.close();
-    let authAction: INinjaAction | undefined;
-    if (authService?.isLoggedIn()) {
-      authAction = {
-        id: 'Logout',
-        title: window.deps.translateString('commandMenu.logout', 'Logout'),
-        mdIcon: 'logout',
-        handler: () => {
-          UI.getLogoutLink()?.click();
-        },
-      };
+    ninja.close();
+    const { actions, keyboardShortcuts, loginAction, logoutAction } = getCommandMenuActions({
+      deps: {
+        getConfig,
+        loadStarterTemplate,
+        changeEditorSettings,
+        changeLayout,
+      },
+    });
+    if (parent === 'Keyboard Shortcuts') {
+      ninja.data = keyboardShortcuts;
+      if (HomeBreadcrumb) {
+        HomeBreadcrumb.innerText = window.deps.translateString(
+          'commandMenu.keyboardShortcuts',
+          'Keyboard Shortcuts',
+        );
+      }
     } else {
-      authAction = {
-        id: 'Login',
-        title: window.deps.translateString('commandMenu.login', 'Login'),
-        mdIcon: 'login',
-        handler: () => {
-          UI.getLoginLink()?.click();
-        },
-      };
+      const authAction = authService?.isLoggedIn() ? logoutAction : loginAction;
+      ninja.data = [...actions, authAction];
+      if (parent) {
+        ninja.setParent(parent);
+      }
+      if (HomeBreadcrumb) {
+        HomeBreadcrumb.innerText = window.deps.translateString('commandMenu.home', 'Home');
+      }
     }
-    ninja.data = adjustTrailingDots([...actions, authAction]);
-    ninja.open();
+    requestAnimationFrame(() => ninja.open());
   };
 
   const onHotkey = async (e: KeyboardEvent) => {
-    const ctrl = (e: KeyboardEvent) => (navigator.platform.match('Mac') ? e.metaKey : e.ctrlKey);
+    const ctrl = (e: KeyboardEvent) => (isMac() ? e.metaKey : e.ctrlKey);
     if (ctrl(e) && e.code === 'KeyK') {
       e.preventDefault();
       // eslint-disable-next-line no-underscore-dangle
@@ -3399,13 +2622,19 @@ const handleCommandMenu = async () => {
       // eslint-disable-next-line no-underscore-dangle
       if (ninja.__visible === false) {
         ninja.focus();
-        requestAnimationFrame(openCommandMenu);
+        requestAnimationFrame(() => openCommandMenu());
       }
     }
   };
 
   eventsManager.addEventListener(window, 'keydown', onHotkey, true);
-  eventsManager.addEventListener(UI.getCommandMenuLink(), 'click', openCommandMenu, true);
+  eventsManager.addEventListener(UI.getCommandMenuLink(), 'click', () => openCommandMenu(), true);
+  eventsManager.addEventListener(
+    UI.getKeyboardShortcutsMenuLink(),
+    'click',
+    () => openCommandMenu('Keyboard Shortcuts'),
+    true,
+  );
 };
 
 const handleLogoLink = () => {
@@ -3619,7 +2848,11 @@ const handleAppMenuProject = () => {
   const menuProjectContainer = UI.getAppMenuProjectScroller();
   const menuProjectButton = UI.getAppMenuProjectButton();
   if (!menuProjectContainer || !menuProjectButton) return;
-  menuProjectContainer.innerHTML = menuProjectHTML; // settingsMenuHTML;
+
+  const html = isMac()
+    ? menuProjectHTML.replace(/<kbd>Ctrl<\/kbd>/g, '<kbd>⌘</kbd>')
+    : menuProjectHTML;
+  menuProjectContainer.innerHTML = html;
   translateElement(menuProjectContainer);
   // adjustFontSize(menuProjectContainer);
 
@@ -3642,7 +2875,11 @@ const handleAppMenuSettings = () => {
   const menuSettingsContainer = UI.getAppMenuSettingsScroller();
   const menuSettingsButton = UI.getAppMenuSettingsButton();
   if (!menuSettingsContainer || !menuSettingsButton) return;
-  menuSettingsContainer.innerHTML = menuSettingsHTML; // settingsMenuHTML;
+
+  const html = isMac()
+    ? menuSettingsHTML.replace(/<kbd>Ctrl<\/kbd>/g, '<kbd>⌘</kbd>')
+    : menuSettingsHTML;
+  menuSettingsContainer.innerHTML = html;
 
   translateElement(menuSettingsContainer);
   adjustFontSize(menuSettingsContainer);
@@ -3666,7 +2903,9 @@ const handleAppMenuHelp = () => {
   const menuHelpContainer = UI.getAppMenuHelpScroller();
   const menuHelpButton = UI.getAppMenuHelpButton();
   if (!menuHelpContainer || !menuHelpButton) return;
-  menuHelpContainer.innerHTML = menuHelpHTML;
+
+  const html = isMac() ? menuHelpHTML.replace(/<kbd>Ctrl<\/kbd>/g, '<kbd>⌘</kbd>') : menuHelpHTML;
+  menuHelpContainer.innerHTML = html;
   translateElement(menuHelpContainer);
   // adjustFontSize(menuHelpContainer);
 
@@ -5340,7 +4579,7 @@ const handleResultZoom = () => {
   const zoomBtn = document.createElement('div');
   zoomBtn.id = 'zoom-button';
   zoomBtn.classList.add('tool-buttons');
-  zoomBtn.title = window.deps.translateString('core.zoom.hint', 'Zoom');
+  zoomBtn.title = window.deps.translateString('core.zoom.hint', 'Zoom') + ' (Ctrl/Cmd + Alt + Z)';
   zoomBtn.style.pointerEvents = 'all'; //  override setting to 'none' on toolspane bar
   zoomBtn.innerHTML = `
   <button class="text">
