+++ src/livecodes/html/app-menu-help.html
@@ -5,6 +5,11 @@
       <span><kbd>Ctrl</kbd><kbd>K</kbd></span>
     </a>
   </li>
+  <li>
+    <a href="#" id="keyboard-shortcuts-menu-link">
+      <span data-i18n="menu.keyboardShortcuts">Keyboard Shortcuts</span>
+    </a>
+  </li>
   <li role="separator"></li>
   <li>
     <a href="https://blog.livecodes.io" target="_blank" rel="noopener" data-i18n="menu.blog"
@@ -46,11 +51,7 @@
   </li>
   <li role="separator"></li>
   <li>
-    <a
-      href="https://github.com/live-codes/livecodes/issues/new/choose"
-      target="_blank"
-      rel="noopener"
-      data-i18n="menu.report"
+    <a href="{{REPO_URL}}/issues/new/choose" target="_blank" rel="noopener" data-i18n="menu.report"
       >Report an issue</a
     >
   </li>
