+++ src/livecodes/i18n/locales/en/translation.ts
@@ -297,6 +297,7 @@ const translation = {
       title: 'Change Theme',
     },
     changeUILanguage: 'Change UI Language',
+    contribute: 'Contribute',
     copy: 'Copy Code',
     copyAsDataUrl: 'Copy Code as Data URL',
     disableAI: 'Disable AI Code Assistant',
@@ -306,6 +307,7 @@ const translation = {
     disableFormatOnSave: 'Disable Format On-Save',
     disableRecoverUnsaved: 'Disable Recover Unsaved',
     disableVim: 'Disable Vim Mode',
+    editorKeyboardShortcuts: 'Editor Keyboard Shortcuts',
     enableAI: 'Enable AI Code Assistant',
     enableAutoSave: 'Enable Auto Save',
     enableAutoUpdate: 'Enable Auto Update',
@@ -313,13 +315,20 @@ const translation = {
     enableFormatOnSave: 'Enable Format On-Save',
     enableRecoverUnsaved: 'Enable Recover Unsaved',
     enableVim: 'Enable Vim Mode',
+    focus: {
+      editor: 'Focus Editor',
+      home: 'Move Focus to Home',
+      outOfEditor: 'Move Focus out of Editor',
+    },
     formatCode: 'Format Code',
     home: 'Home',
     horizontalLayout: 'Horizontal Layout',
+    keyboardShortcuts: 'Keyboard Shortcuts',
     login: 'Login',
     logout: 'Logout',
     moveToParent: 'move to parent',
     placeholder: 'Type a command or search...',
+    processors: 'Processors',
     responsiveLayout: 'Responsive Layout',
     run: 'Run',
     saveAsFork: 'Save as a Fork (New Project)',
@@ -342,12 +351,19 @@ const translation = {
       style: 'Show Style Editor',
       tests: 'Toggle Tests',
       title: 'Show …',
+      zoom: 'Toggle Result Zoom',
     },
     starterTemplates: 'Starter Templates',
+    sync: 'Sync (beta) …',
     template: 'Template',
+    theme: {
+      color: 'Set Theme Color',
+      defaultColor: 'Set Default Theme Color',
+    },
     toClose: 'to close',
     toNavigate: 'to navigate',
     toSelect: 'to select',
+    toggle: 'Toggle: ',
     verticalLayout: 'Vertical Layout',
   },
   core: {
@@ -701,6 +717,7 @@ const translation = {
     formatOnsave: 'Format On-save',
     getstart: 'Getting Started',
     import: 'Import …',
+    keyboardShortcuts: 'Keyboard Shortcuts',
     layout: 'Vertical Layout',
     license: 'License',
     login: 'Login',
